requirements.txt
# All thinks what the bot need
python-dotenv
py-cord
asyncio